+++ app/src/main/java/page/x/TLB/algoritmos/substituicao/LRU.java
@@ -18,7 +18,7 @@ public class LRU implements AlgoritmoSubstituicaoI {
     }
 
     @Override
-    public Integer mapearPagina(Integer vpn) throws MissInterruption {
+    public Long mapearPagina(Long vpn) throws MissInterruption {
         for (TlbEntry entryAtual : this.entries) {
             if (entryAtual.getVirtualPageNumber().equals(vpn)) {
                 this.execucaoAlgoritmo(entryAtual);
@@ -34,7 +34,7 @@ public class LRU implements AlgoritmoSubstituicaoI {
     }    
 
     @Override
-    public void addPaginaMapeada(Integer vpn, Integer pfn) {
+    public void addPaginaMapeada(Long vpn, Long pfn) {
         TlbEntry tlbEntry = new TlbEntry(vpn, pfn);
         if (quantidadeEntries == this.entries.size()) {
             this.entries.removeFirst();
