+++ app/src/test/java/page/x/TestLFU.java
@@ -32,7 +32,7 @@ class TestLFU {
         this.tlb1 = new TLB(lfu1);
         this.entries = new TlbEntry[QTD_PAIR];
         for (int i = 0; i < QTD_PAIR; i++) {
-            this.entries[i] = new TlbEntry(i, i*2);
+            this.entries[i] = new TlbEntry((long) i, (long) i*2);
         }
     }
 
@@ -41,18 +41,18 @@ class TestLFU {
     void testandoLogicaDaLFU() throws MissInterruption {
         this.preencheTLB(tlb1, entries, SIZE_LFU);
         for (int i = 0; i < 3; i++) {
-            this.tlb1.mapearPagina(i);
-            this.tlb1.mapearPagina(i + 1);
+            this.tlb1.mapearPagina((long) i);
+            this.tlb1.mapearPagina((long) (i + 1));
         }
-        this.tlb1.addPaginaMapeada(50, 100);
+        this.tlb1.addPaginaMapeada(50L, 100L);
         assertAll(
-            () -> assertThrows(MissInterruption.class, () -> this.tlb1.mapearPagina(4)),
+            () -> assertThrows(MissInterruption.class, () -> this.tlb1.mapearPagina(4L)),
             () -> assertEquals(6, tlb1.getHit()),
             () -> assertEquals(1, tlb1.getMiss())
         );
-        this.tlb1.addPaginaMapeada(51, 101);
+        this.tlb1.addPaginaMapeada(51L, 101L);
         assertAll(
-            () -> assertThrows(MissInterruption.class, () -> this.tlb1.mapearPagina(50)),
+            () -> assertThrows(MissInterruption.class, () -> this.tlb1.mapearPagina(50L)),
             () -> assertEquals(6, tlb1.getHit()),
             () -> assertEquals(2, tlb1.getMiss())
         );
